bone-iio-helper: fix driver initialization order.
When you declared in the main dts the bone-iio-helper node and
driver is compiled statically into the kernel,
the helper driver is initialized before the ti_am335x_adc driver.
This generate a bug during boot initialization
because bone-iio-helper depends by ti_am335x_adc.

Fix it, changing the order of initialization of the bone-iio-helper
and placed among the last drivers to be init with late_initcall().

Signed-off-by: Matteo Facchinetti <matteo.facchinetti@sirius-es.it>